# ga4gh/templates/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GA4GH Reference Server",
        "version": "0.2.0",
        "title": "GA4GH Reference Server",
        "contact": {
            "email": "someone@ucsc.edu"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:8000",
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/references/{id}": {
            "get": {
                "tags": [
                    "references"
                ],
                "description": "Returns references for a given ID",
                "summary": "Find references by ID",
                "operationId": "getReference",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "reference response",
                        "schema": {
                            "$ref": "#/definitions/Reference"
                        }
                    }
                }
            },
            "parameters": [{
                "name": "id",
                "in": "path",
                "description": "ID of the reference",
                "required": true,
                "type": "string"
            }]
        },
        "/referencesets/{id}": {
            "get": {
                "tags": [
                    "referencesets"
                ],
                "description": "Returns reference sets for a given ID",
                "summary": "Find reference sets by ID",
                "operationId": "getReferenceSet",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "reference sets response",
                        "schema": {
                            "$ref": "#/definitions/ReferenceSet"
                        }
                    }
                }
            },
            "parameters": [{
                "name": "id",
                "in": "path",
                "description": "ID of the reference set",
                "required": true,
                "type": "string"
            }]
        },
        "/references/{id}/bases": {
            "get": {
                "tags": [
                    "referencebases"
                ],
                "description": "Returns reference bases for a given ID",
                "summary": "Find reference bases by ID",
                "operationId": "getReferencebases",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "reference bases response",
                        "schema": {
                            "$ref": "#/definitions/ListReferenceBasesResponse"
                        }
                    }
                }
            },
            "parameters": [{
                "name": "id",
                "in": "path",
                "description": "ID of the reference bases",
                "required": true,
                "type": "string"
            }]
        },
        "/callsets/search": {
            "post": {
                "tags": [
                    "callsets"
                ],
                "summary": "Search for callsets",
                "description": "",
                "operationId": "searchCallsets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Search request",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SearchCallSetsRequest"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SearchCallSetsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "The specified Call Set ID was not found"
                    }
                }
            }
        },
        "/readgroupsets/search": {
            "post": {
                "tags": [
                    "readgroupset"
                ],
                "summary": "Search for readgroupsets",
                "description": "",
                "operationId": "searchReadGroupSets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Search request",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SearchReadGroupSetsRequest"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SearchReadGroupSetsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "The specified Read Group Set ID was not found"
                    }
                }
            }
        },
        "/reads/search": {
            "post": {
                "tags": [
                    "reads"
                ],
                "summary": "Search for reads",
                "description": "",
                "operationId": "searchReads",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Search request",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SearchReadsRequest"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SearchReadsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "The specified Read ID was not found"
                    }
                }
            }
        },
        "/referencesets/search": {
            "post": {
                "tags": [
                    "referenceset"
                ],
                "summary": "Search for referencesets",
                "description": "",
                "operationId": "searchReferenceSets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Search request",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SearchReferenceSetsRequest"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SearchReferenceSetsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "The specified Reference Set ID was not found"
                    }
                }
            }
        },
        "/references/search": {
            "post": {
                "tags": [
                    "reference"
                ],
                "summary": "Search for references",
                "description": "",
                "operationId": "searchReferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Search request",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SearchReferencesRequest"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SearchReferencesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "The specified Reference ID was not found"
                    }
                }
            }
        },
        "/variantsets/search": {
            "post": {
                "tags": [
                    "variantset"
                ],
                "summary": "Search for variantsets",
                "description": "",
                "operationId": "searchVariantSets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Search request",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SearchVariantSetsRequest"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SearchVariantSetsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "The specified Variant Set ID was not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Reference": {
            "type": "object",
            "required": [
                "id",
                "length",
                "md5checksum",
                "name",
                "sourceURI",
                "sourceAccessions",
                "isDerived"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "length": {
                    "type": "integer"
                },
                "md5checksum": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sourceURI": {
                    "type": "string"
                },
                "sourceAccessions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isDerived": {
                    "type": "boolean"
                }
            }
        },
...
        "ListReferenceBasesResponse": {
            "type": "object",
            "required": [
                "offset",
                "sequence"
            ],
            "properties": {
                "offset": {
                    "type": "integer"
                },
                "sequence": {
                    "type": "string"
                },
                "nextPageToken": {
                    "type": "string"
                }
            }
        },
...
        "SearchCallSetsRequest": {
            "type": "object",
            "required": [
                "variantSetId"
            ],
            "properties": {
                "variantSetId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pageSize": {
                    "type": "integer"
                },
                "pageToken": {
                    "type": "string"
                }
            }
        },
...
    },
...
}